backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.30
alembic>=1.14.0
psycopg[binary]>=3.2.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
redis>=5.1.0
python-multipart>=0.0.20
email-validator>=2.2.0
google-genai>=1.0.0
